pyproject.toml
[project]
name = "provenance_support"
version = "1.2.15"
dependencies = ['keyrings.efile',
    'postgresql-access',
    'pyyaml']
#    'provenance_track']
requires-python= ">= 3.10"
readme = "README.md"
authors = [
    {name = "Gerard"},
    {email = "gweatherby@uchc.edu"}
]
dynamic = ["license"]

[project.optional-dependencies]

[project.scripts]
createtable = "ptsupport.createtable:main"
modelschema = "ptsupport.modelschema:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

